+++ CHANGES.md
@@ -17,11 +17,11 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
 - m3api now has a stable interface policy, see [README.md § stability](README.md#Stability).
   Internal breaking changes, such as the one above, do not require bumping the major version number.
 - The new `authorization` request option can be used to set an `Authorization:` request header.
-  This will be used by the upcoming m3api-oauth package to support OAuth 2.0.
+  This will be used by the upcoming m3api-oauth2 package to support OAuth 2.0.
 - The `apiUrl` member of a `Session` is now public.
-  This will be used by the upcoming m3api-oauth package.
+  This will be used by the upcoming m3api-oauth2 package.
 - A new protected method, `getUserAgent()`, can be used to get the user agent for a request.
-  This will be used by the upcoming m3api-oauth package.
+  This will be used by the upcoming m3api-oauth2 package.
 - The `dropTruncatedResultWarning` didn’t work properly
   when using the default `errorformat` on current MediaWiki versions;
   this has been fixed.