+++ CHANGES.md
@@ -5,7 +5,7 @@ This file records the changes in each m3api release.
 The annotated tag (and GitHub release) for each version also lists the changes,
 but this file may sometimes contain later improvements (e.g. typo fixes).
 
-## next (not yet released)
+## v0.7.2 (2022-12-04)
 
 - Internal Breaking Change:
   The signature of the `internalGet()` and `internalPost()` methods has changed;